See merge request karaboDevices/dsscDevices!8

+++ DsscProcessor/src/DsscProcessor.cc
@@ -464,7 +464,7 @@ namespace karabo {
       const uint16_t* data_ptr = data.getData<uint16_t>();
       size_t data_size = data.size();
 
-      const uint32_t* cellId_ptr = cellId.getData<uint32_t>();
+      const uint16_t* cellId_ptr = cellId.getData<uint16_t>();
       size_t cellId_size = cellId.size();
 
       const unsigned long long* trainId_ptr = trainId.getData<unsigned long long>();
